Extention folder for the UIColor constants added
